imp: migrate JSON for feemarket
