+++ x/feemarket/keeper/migrations_test.go
@@ -4,6 +4,7 @@ import (
 	"math/big"
 
 	"github.com/tharsis/ethermint/x/feemarket/keeper"
+	"github.com/tharsis/ethermint/x/feemarket/v0_12"
 )
 
 func (suite *KeeperTestSuite) TestMigration1To2() {
@@ -11,7 +12,7 @@ func (suite *KeeperTestSuite) TestMigration1To2() {
 	storeKey := suite.app.GetKey("feemarket")
 	store := suite.ctx.KVStore(storeKey)
 	baseFee := big.NewInt(1000)
-	store.Set(keeper.KeyPrefixBaseFeeV1, baseFee.Bytes())
+	store.Set(v0_12.KeyPrefixBaseFeeV1, baseFee.Bytes())
 	m := keeper.NewMigrator(suite.app.FeeMarketKeeper)
 	suite.Require().NoError(m.Migrate1to2(suite.ctx))
 	suite.Require().Equal(baseFee, suite.app.FeeMarketKeeper.GetBaseFee(suite.ctx))
